requirements.txt
discord.py==2.3.2
rustplus==5.5.2
python-dotenv==1.0.0
aiosqlite==0.19.0
Pillow==10.2.0
aiohttp==3.9.3
twilio==8.1.0
